+++ akka-sample-multi-node-scala/build.sbt
@@ -7,6 +7,7 @@ val `akka-sample-multi-node-scala` = project
   .in(file("."))
   .settings(multiJvmSettings: _*)
   .settings(
+    organization := "com.typesafe.akka.samples",
     scalaVersion := "2.12.1",
     libraryDependencies ++= Seq(
       "com.typesafe.akka" %% "akka-actor" % akkaVersion,
