--- akka-sample-persistence-java-lambda/.gitignore
@@ -1,17 +1,0 @@
-*#
-*.iml
-*.ipr
-*.iws
-*.pyc
-*.tm.epoch
-*.vim
-*-shim.sbt
-.idea/
-/project/plugins/project
-project/boot
-target/
-/logs
-.cache
-.classpath
-.project
-.settings
