--- akka-sample-persistence-java-lambda/LICENSE
@@ -1,10 +1,0 @@
-Activator Template by Lightbend
-
-Licensed under Public Domain (CC0)
-
-To the extent possible under law, the person who associated CC0 with
-this Activator Tempate has waived all copyright and related or neighboring
-rights to this Activator Template.
-
-You should have received a copy of the CC0 legalcode along with this
-work.  If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
