--- akka-sample-persistence-java-lambda/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-persistence-java-lambda
-title=Akka Persistence Samples in Java with Lambdas
-description=Akka Persistence Samples in Java with Lambdas
-tags=akka,java,java8,sample,persistence,sample
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
