--- akka-sample-persistence-java-lambda/project/build.properties
@@ -1,1 +1,0 @@
-sbt.version=0.13.7
