--- akka-sample-persistence-java-lambda/src/main/resources/application.conf
@@ -1,9 +1,0 @@
-akka.persistence.journal.plugin = "akka.persistence.journal.leveldb"
-akka.persistence.snapshot-store.plugin = "akka.persistence.snapshot-store.local"
-
-akka.persistence.journal.leveldb.dir = "target/example/journal"
-akka.persistence.snapshot-store.local.dir = "target/example/snapshots"
-
-# DO NOT USE THIS IN PRODUCTION !!!
-# See also https://github.com/typesafehub/activator/issues/287
-akka.persistence.journal.leveldb.native = false
