+++ akka-sample-persistence-java/LICENSE
@@ -1,4 +1,4 @@
-Activator Template by Lightbend
+Akka sample by Lightbend
 
 Licensed under Public Domain (CC0)
