--- akka-sample-persistence-java/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-persistence-java
-title=Akka Persistence Samples with Java
-description=Akka Persistence Samples with Java
-tags=akka,persistence,java,sample
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
