+++ akka-sample-persistence-java/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("com.dwijnand" % "sbt-dynver" % "1.1.1")
