+++ akka-sample-persistence-java/src/main/java/sample/persistence/ViewExample.java
