--- akka-sample-persistence-scala/activator.properties
@@ -1,7 +1,0 @@
-name=akka-sample-persistence-scala
-title=Akka Persistence Samples with Scala
-description=Akka Persistence Samples with Scala
-tags=akka,persistence,scala,sample
-authorName=Akka Team
-authorLink=http://akka.io/
-sourceLink=https://github.com/akka/akka
