+++ akka-sample-persistence-scala/build.sbt
@@ -1,12 +1,10 @@
+organization := "com.typesafe.akka.samples"
 name := "akka-sample-persistence-scala"
 
-version := "2.4-SNAPSHOT"
-
-scalaVersion := "2.11.7"
-
+scalaVersion := "2.12.1"
 libraryDependencies ++= Seq(
-  "com.typesafe.akka" %% "akka-actor" % "2.4-SNAPSHOT",
-  "com.typesafe.akka" %% "akka-persistence" % "2.4-SNAPSHOT",
+  "com.typesafe.akka" %% "akka-actor" % "2.4.14",
+  "com.typesafe.akka" %% "akka-persistence" % "2.4.14",
   "org.iq80.leveldb"            % "leveldb"          % "0.7",
   "org.fusesource.leveldbjni"   % "leveldbjni-all"   % "1.8"
 )
